Bugfix: don't throw an error when the Id or Rev fields aren't found on the object while serializing them.
Throwing an error in that case introduced an interesting bug in the C# wrapper. Our default converter's serializer settings are set to ignore null values. C# classes can have null values, so when the converter initially turns the given doc class into a JToken it can potentially completely remove those fields when they're null. That would cause the converter to throw the error because it thinks the fields don't exist on the class that C# gave it.